[AI-68] daemon: WIP extract ClaudeLauncher (next: wire through orchestrator)
