+++ src/Kapacitor.Daemon/Services/ClaudeLauncher.cs
@@ -1,0 +1,400 @@
+using System.Text.Json;
+using System.Text.Json.Nodes;
+using kapacitor;
+using kapacitor.Commands;
+using Microsoft.Extensions.Logging;
+
+namespace kapacitor.Daemon.Services;
+
+internal sealed partial class ClaudeLauncher(
+        DaemonConfig            config,
+        ILogger<ClaudeLauncher> logger
+    ) : IHostedAgentLauncher {
+
+    public string Vendor  => "claude";
+    public string CliPath => config.ClaudePath;
+
+    static readonly Lock                  TrustWriteLock   = new();
+    static readonly JsonSerializerOptions IndentedJsonOpts = new() { WriteIndented = true };
+
+    public void Prepare(LauncherContext ctx) {
+        // Overlay .claude/ local settings from source repo into worktree.
+        try {
+            var sourceClaudeDir = Path.Combine(ctx.SourceRepoPath, ".claude");
+            var destClaudeDir   = Path.Combine(ctx.Worktree.Path, ".claude");
+
+            if (Directory.Exists(sourceClaudeDir)) {
+                OverlayDirectory(sourceClaudeDir, destClaudeDir);
+            }
+
+            SymlinkClaudeProjectDir(ctx.SourceRepoPath, ctx.Worktree.Path);
+        } catch (Exception ex) {
+            LogOverlayFailed(ex, ctx.AgentId);
+        }
+
+        try {
+            WriteMcpConfig(ctx.SourceRepoPath, ctx.Worktree.Path);
+        } catch (Exception ex) {
+            LogMcpConfigFailed(ex, ctx.AgentId);
+        }
+
+        try {
+            TrustWorktreeInClaudeConfig(ctx.Worktree.Path);
+        } catch (Exception ex) {
+            LogTrustWorktreeFailed(ex, ctx.AgentId);
+        }
+
+        try {
+            if (ctx.Tools is { Length: > 0 }) {
+                MergeToolPermissions(ctx.Worktree.Path, ctx.Tools);
+            }
+        } catch (Exception ex) {
+            LogToolPermissionsFailed(ex, ctx.AgentId);
+        }
+    }
+
+    public LaunchArgs BuildArgs(LauncherContext ctx) {
+        var args = new List<string>();
+        string? mcpConfigPath = null;
+
+        if (ctx.IsReview && ctx.ReviewLaunch is { } launch) {
+            mcpConfigPath = launch.McpConfigPath;
+
+            args.Add("--mcp-config");
+            args.Add(launch.McpConfigPath);
+            args.Add("--system-prompt");
+            args.Add(launch.SystemPrompt);
+
+            if (!string.IsNullOrEmpty(ctx.Model)) {
+                args.Add("--model");
+                args.Add(ctx.Model);
+            }
+        } else {
+            if (!string.IsNullOrEmpty(ctx.Effort)) {
+                args.Add("--effort");
+                args.Add(ctx.Effort);
+            }
+
+            if (!string.IsNullOrEmpty(ctx.Model)) {
+                args.Add("--model");
+                args.Add(ctx.Model);
+            }
+
+            if (!string.IsNullOrEmpty(ctx.Prompt)) {
+                args.Add("--");
+                args.Add(ctx.Prompt);
+            }
+        }
+
+        return new LaunchArgs(args.ToArray(), mcpConfigPath);
+    }
+
+    public void Cleanup(AgentInstance agent) {
+        try { RemoveClaudeProjectSymlink(agent.Worktree.Path); } catch (Exception ex) {
+            LogCleanupSymlinkFailed(ex, agent.Id);
+        }
+
+        if (agent.McpConfigPath is not null) {
+            try { File.Delete(agent.McpConfigPath); } catch (Exception ex) {
+                LogCleanupMcpConfigFailed(ex, agent.Id);
+            }
+        }
+    }
+
+    // === Moved verbatim from AgentOrchestrator.cs ===
+
+    /// <summary>
+    /// Merges dialog-selected tool names into the worktree's .claude/settings.local.json
+    /// permissions.allow array. Existing granular rules (e.g. "Bash(git:*)") are preserved;
+    /// dialog selections add broad tool-level entries (e.g. "Bash").
+    /// </summary>
+    internal static void MergeToolPermissions(string worktreePath, string[] tools) {
+        var settingsPath = Path.Combine(worktreePath, ".claude", "settings.local.json");
+
+        JsonNode? root = null;
+
+        if (File.Exists(settingsPath)) {
+            try {
+                root = JsonNode.Parse(File.ReadAllText(settingsPath));
+            } catch {
+                // Malformed JSON — start fresh
+            }
+        }
+
+        if (root is not JsonObject rootObj) {
+            rootObj = [];
+        }
+
+        if (rootObj["permissions"] is not JsonObject permissions) {
+            permissions            = [];
+            rootObj["permissions"] = permissions;
+        }
+
+        if (permissions["allow"] is not JsonArray allow) {
+            allow                = [];
+            permissions["allow"] = allow;
+        }
+
+        var existing = new HashSet<string>(
+            allow.Select(n => (n as JsonValue)?.TryGetValue<string>(out var s) == true ? s : null)
+                .Where(s => s != null)!
+        );
+
+        foreach (var tool in tools) {
+            if (!existing.Contains(tool)) {
+                allow.Add((JsonNode)JsonValue.Create(tool)!);
+                existing.Add(tool);
+            }
+        }
+
+        Directory.CreateDirectory(Path.GetDirectoryName(settingsPath)!);
+
+        File.WriteAllText(settingsPath, rootObj.ToJsonString(IndentedJsonOpts));
+    }
+
+    /// <summary>
+    /// Copies files from <paramref name="source"/> into <paramref name="dest"/>,
+    /// creating directories as needed but never overwriting existing files.
+    /// This preserves git-tracked files while adding gitignored local settings.
+    /// </summary>
+    static void OverlayDirectory(string source, string dest) {
+        Directory.CreateDirectory(dest);
+
+        foreach (var file in Directory.GetFiles(source)) {
+            var destFile = Path.Combine(dest, Path.GetFileName(file));
+
+            if (!File.Exists(destFile)) {
+                File.Copy(file, destFile);
+            }
+        }
+
+        foreach (var dir in Directory.GetDirectories(source)) {
+            var destDir = Path.Combine(dest, Path.GetFileName(dir));
+            OverlayDirectory(dir, destDir);
+        }
+    }
+
+    /// <summary>
+    /// Creates a symlink at ~/.claude/projects/{worktree-path-hash} pointing to
+    /// ~/.claude/projects/{source-path-hash} so project-level permissions, settings,
+    /// and memory are shared with the hosted agent.
+    /// </summary>
+    static void SymlinkClaudeProjectDir(string sourceRepoPath, string worktreePath) {
+        if (!Directory.Exists(ClaudePaths.Projects)) {
+            return;
+        }
+
+        var sourceProjDir = ClaudePaths.ProjectDir(sourceRepoPath);
+
+        if (!Directory.Exists(sourceProjDir)) {
+            return;
+        }
+
+        var worktreeProjDir = ClaudePaths.ProjectDir(worktreePath);
+
+        // Don't clobber an existing directory or symlink
+        if (Path.Exists(worktreeProjDir)) {
+            return;
+        }
+
+        Directory.CreateSymbolicLink(worktreeProjDir, sourceProjDir);
+    }
+
+    /// <summary>
+    /// Removes the ~/.claude/projects/{worktree-path-hash} symlink if it exists.
+    /// Only removes symlinks, never real directories.
+    /// </summary>
+    static void RemoveClaudeProjectSymlink(string worktreePath) {
+        var info = new DirectoryInfo(ClaudePaths.ProjectDir(worktreePath));
+
+        if (info is { Exists: true, LinkTarget: not null }) {
+            info.Delete();
+        }
+    }
+
+    static void TrustWorktreeInClaudeConfig(string worktreePath) {
+        // Serialize against concurrent agent launches. ~/.claude.json is shared
+        // across the whole user and {worktree}/.claude/settings.local.json is
+        // touched by MergeToolPermissions on the same path; interleaved reads and
+        // writes could drop updates or write truncated JSON.
+        lock (TrustWriteLock) {
+            // Workspace trust (the "Do you trust the files in this folder?" dialog)
+            // is stored globally in ~/.claude.json under projects[path]. On a fresh
+            // machine the file may not exist yet — create a minimal object so trust
+            // is always persisted.
+            var claudeJsonPath = Path.Combine(PathHelpers.HomeDirectory, ".claude.json");
+            var root           = LoadJsonObject(claudeJsonPath);
+
+            if (root["projects"] is not JsonObject projects) {
+                projects         = [];
+                root["projects"] = projects;
+            }
+
+            if (projects[worktreePath] is not JsonObject entry) {
+                entry                  = [];
+                projects[worktreePath] = entry;
+            }
+
+            var alreadyTrusted = entry["hasTrustDialogAccepted"] is JsonValue v && v.TryGetValue<bool>(out var b) && b;
+
+            if (!alreadyTrusted) {
+                entry["hasTrustDialogAccepted"] = true;
+                WriteJsonAtomic(claudeJsonPath, root);
+            }
+
+            // MCP server approval (the "New MCP server found in .mcp.json" dialog)
+            // is stored per-project in {worktree}/.claude/settings.local.json, NOT in
+            // ~/.claude.json. Claude requires BOTH fields: the blanket flag AND each
+            // server listed by name (the blanket flag alone still shows a per-server
+            // discovery prompt on first encounter).
+            var serverNames = ReadMcpJsonServerNames(worktreePath);
+
+            if (serverNames.Count == 0) return;
+
+            var settingsDir  = Path.Combine(worktreePath, ".claude");
+            var settingsPath = Path.Combine(settingsDir, "settings.local.json");
+
+            Directory.CreateDirectory(settingsDir);
+
+            var settings = LoadJsonObject(settingsPath);
+            var sDirty   = false;
+
+            var allEnabled = settings["enableAllProjectMcpServers"] is JsonValue ev
+             && ev.TryGetValue<bool>(out var eb)
+             && eb;
+
+            if (!allEnabled) {
+                settings["enableAllProjectMcpServers"] = true;
+                sDirty                                 = true;
+            }
+
+            var known = new HashSet<string>();
+
+            if (settings["enabledMcpjsonServers"] is JsonArray arr) {
+                foreach (var item in arr) {
+                    if (item is JsonValue jv && jv.TryGetValue<string>(out var s)) {
+                        known.Add(s);
+                    }
+                }
+            }
+
+            if (serverNames.Any(known.Add)) {
+                // Rebuild the array via JSON parsing to avoid AOT issues with
+                // JsonArray.Add(string) / JsonValue.Create<string>.
+                var json = "[" + string.Join(",", known.Select(n => $"\"{JsonEncodedText.Encode(n)}\"")) + "]";
+                settings["enabledMcpjsonServers"] = JsonNode.Parse(json);
+                sDirty                            = true;
+            }
+
+            if (sDirty) {
+                WriteJsonAtomic(settingsPath, settings);
+            }
+        }
+    }
+
+    /// <summary>
+    /// Loads a JSON object from disk, tolerating missing files and non-object roots
+    /// by returning a fresh <see cref="JsonObject"/>. Ensures pre-trust logic never
+    /// throws on an unexpected config shape and reintroduces the launch hang.
+    /// </summary>
+    static JsonObject LoadJsonObject(string path) {
+        if (!File.Exists(path)) return new JsonObject();
+
+        try {
+            return JsonNode.Parse(File.ReadAllText(path)) as JsonObject ?? new JsonObject();
+        } catch {
+            return new JsonObject();
+        }
+    }
+
+    /// <summary>
+    /// Writes JSON to <paramref name="path"/> via a sibling temp file + atomic
+    /// rename (POSIX <c>rename(2)</c> / Win32 <c>MoveFileEx</c> with replace).
+    /// A crash or concurrent reader never sees a truncated file — either the
+    /// previous contents or the new contents, never a partial write.
+    /// </summary>
+    static void WriteJsonAtomic(string path, JsonNode root) {
+        var tmp = path + ".tmp-" + Environment.ProcessId + "-" + Guid.NewGuid().ToString("N");
+        File.WriteAllText(tmp, root.ToJsonString(IndentedJsonOpts));
+
+        try {
+            File.Move(tmp, path, overwrite: true);
+        } catch {
+            try { File.Delete(tmp); } catch {
+                /* best-effort */
+            }
+
+            throw;
+        }
+    }
+
+    static List<string> ReadMcpJsonServerNames(string worktreePath) {
+        var mcpJsonPath = Path.Combine(worktreePath, ".mcp.json");
+
+        if (!File.Exists(mcpJsonPath)) return [];
+
+        try {
+            var parsed  = JsonNode.Parse(File.ReadAllText(mcpJsonPath));
+            var servers = parsed?["mcpServers"]?.AsObject();
+
+            return servers is null ? [] : [..servers.Select(kv => kv.Key)];
+        } catch {
+            return [];
+        }
+    }
+
+    static void WriteMcpConfig(string sourceRepoPath, string worktreePath) {
+        var claudeJsonPath = Path.Combine(PathHelpers.HomeDirectory, ".claude.json");
+
+        if (!File.Exists(claudeJsonPath)) return;
+
+        var root    = JsonNode.Parse(File.ReadAllText(claudeJsonPath));
+        var servers = root?["projects"]?[sourceRepoPath]?["mcpServers"]?.AsObject();
+
+        if (servers is null || servers.Count == 0) return;
+
+        var mcpJsonPath = Path.Combine(worktreePath, ".mcp.json");
+
+        // Read existing .mcp.json if present (e.g. committed to git)
+        JsonObject merged;
+
+        if (File.Exists(mcpJsonPath)) {
+            var existing = JsonNode.Parse(File.ReadAllText(mcpJsonPath));
+            merged = existing?["mcpServers"]?.AsObject() ?? new JsonObject();
+        } else {
+            merged = new JsonObject();
+        }
+
+        // Add servers from ~/.claude.json (don't overwrite repo-committed ones)
+        foreach (var (name, value) in servers) {
+            if (!merged.ContainsKey(name) && value is not null) {
+                var clone = value.DeepClone().AsObject();
+                clone.Remove("env");
+                merged[name] = clone;
+            }
+        }
+
+        var wrapper = new JsonObject { ["mcpServers"] = merged };
+        File.WriteAllText(mcpJsonPath, wrapper.ToJsonString(IndentedJsonOpts));
+    }
+
+    // === LoggerMessage source-generated methods ===
+
+    [LoggerMessage(Level = LogLevel.Warning, Message = "Failed to overlay .claude settings for agent {AgentId} (continuing)")]
+    partial void LogOverlayFailed(Exception ex, string agentId);
+
+    [LoggerMessage(Level = LogLevel.Warning, Message = "Failed to write .mcp.json for agent {AgentId} (continuing)")]
+    partial void LogMcpConfigFailed(Exception ex, string agentId);
+
+    [LoggerMessage(Level = LogLevel.Warning, Message = "Failed to pre-trust worktree for agent {AgentId} (continuing)")]
+    partial void LogTrustWorktreeFailed(Exception ex, string agentId);
+
+    [LoggerMessage(Level = LogLevel.Warning, Message = "Failed to merge tool permissions for agent {AgentId} (continuing)")]
+    partial void LogToolPermissionsFailed(Exception ex, string agentId);
+
+    [LoggerMessage(Level = LogLevel.Warning, Message = "Failed to remove claude project symlink for agent {AgentId}")]
+    partial void LogCleanupSymlinkFailed(Exception ex, string agentId);
+
+    [LoggerMessage(Level = LogLevel.Warning, Message = "Failed to remove mcp config for agent {AgentId}")]
+    partial void LogCleanupMcpConfigFailed(Exception ex, string agentId);
+}
